Move auth0 stuff into separate app
This moves some of the auth0 stuff into a separate app. It's going to get a
little more complicated when we add a middleware and it makes sense to split it
out so we can hone it into a library.
