+++ standup/settings.py
@@ -53,6 +53,7 @@ INSTALLED_APPS = (
     'raven.contrib.django.raven_compat',
 
     'standup.api',
+    'standup.auth0',
     'standup.status',
 )
